refactor(algorithm): replaced dictionary instantiations with algorithms
